recipe_backend/requirements.txt
celery==4.2.1
Django==2.0.3
django-cors-headers==2.3.0
django-crispy-forms==1.7.2
django-debug-toolbar==1.9.1
django-extensions==2.0.7
django-filter==2.0.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
factory-boy==2.11.1
Faker==0.8.17
psycopg2==2.7.4
PyJWT==1.6.4
python-dateutil==2.7.3
pytz==2018.4